+++ controllers/tests/controller_server/csi_controller_server_test.py
@@ -34,7 +34,7 @@ from controllers.tests.controller_server.common import (mock_get_agent,
 from controllers.tests.utils import ProtoBufMock
 
 CONTROLLER_SERVER_PATH = "controllers.servers.csi.csi_controller_server"
-TEST_SERVER_PATH = "controllers.servers.utils"
+CONTROLLER_SERVER_UTILS_PATH = "controllers.servers.utils"
 
 
 class BaseControllerSetUp(unittest.TestCase):
@@ -1003,7 +1003,7 @@ class TestPublishVolume(BaseControllerSetUp, CommonControllerTest):
         self.mock_initiators = {'fc_wwns': [self.fc_port],
                                 'iscsi_iqns': [self.iqn],
                                 'nvme_nqns': []}
-        mock_get_node_initiators(self, TEST_SERVER_PATH)
+        mock_get_node_initiators(self, CONTROLLER_SERVER_UTILS_PATH)
 
     def test_publish_volume_success(self):
         self.servicer.ControllerPublishVolume(self.request, self.context)
@@ -1152,7 +1152,7 @@ class TestUnpublishVolume(BaseControllerSetUp, CommonControllerTest):
         self.mock_initiators = {'fc_wwns': [self.fc_port],
                                 'iscsi_iqns': [self.iqn],
                                 'nvme_nqns': []}
-        mock_get_node_initiators(self, TEST_SERVER_PATH)
+        mock_get_node_initiators(self, CONTROLLER_SERVER_UTILS_PATH)
 
     def test_unpublish_volume_success(self):
         self.servicer.ControllerUnpublishVolume(self.request, self.context)
